refactor: 패키지명 entity로 변경 및 불필요한 import문 삭제
