+++ src/main/java/org/sopt/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package org.sopt.user.repository;
 
-import org.sopt.user.domain.User;
+import org.sopt.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
